Fixed bug that combat strength wasn't updated in the UI.
